requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.24.1
azure-eventhub==5.10.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Deprecated==1.2.13
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
